Extract log component view into own template/view
