+++ lib/dsa_web/templates/log/base_roll.html.leex
@@ -1,0 +1,23 @@
+<div class='mr-3'>
+  <span class='inline-block text-xs font-semibold leading-6 px-1 rounded text-blue-500 bg-blue-50 border border-blue-200'><%= @entry.character %></span>
+  <span class='mx-1 lg:mx-2'>»</span>
+  <span class='inline-block text-xs font-semibold leading-6 px-1 rounded bg-yellow-50 text-yellow-600 border border-yellow-200'>
+    <%= trait(@entry.x1) %><%=
+      cond do
+        @entry.x3 == 0 -> nil
+        @entry.x3 > 0 -> "+#{@entry.x3}"
+        true -> "-#{abs(@entry.x3)}"
+      end
+    %>
+  </span>
+</div>
+
+<div class='flex'>
+  <%= if input_value(@f, :dice) do %>
+    <%= tag_helper :dice, @entry.x4 %>
+    <%= if @entry.x4 == 1 || @entry.x4 == 20, do: tag_helper(:dice, @entry.x5) %>
+    <%= if input_value(@f, :result), do: separator() %>
+  <% end %>
+  <%= if input_value(@f, :result), do: result_tag(:trait, @entry.x12) %>
+</div>
+
