+++ lib/dsa_web/templates/log/empty.html.leex
@@ -1,0 +1,1 @@
+Bei Phex! Noch sind keine Würfel gerollt!
