+++ lib/dsa_web/templates/log/quick_roll.html.leex
@@ -1,0 +1,28 @@
+<div class='mr-3'>
+  <span class='inline-block text-xs font-semibold leading-6 px-1 rounded text-blue-500 bg-blue-50 border border-blue-200'><%= @entry.character %></span>
+  <span class='mx-1 lg:mx-2'>»</span>
+  <span class='inline-block text-xs font-semibold leading-6 px-1 rounded bg-yellow-50 text-yellow-600 border border-yellow-200'>
+    <%= if @entry.x11 > 1, do: @entry.x11 %>W<%= @entry.x10 %>
+    <%=
+      cond do
+        @entry.x9 == 0 -> nil
+        @entry.x9 > 0 -> "+ #{@entry.x9}"
+        true -> "- #{abs(@entry.x9)}"
+      end
+    %>
+  </span>
+</div>
+<div class='flex'>
+  <%= if input_value(@f, :dice) do %>
+    <%= tag_helper :dice, @entry.x1 %>
+    <%= if @entry.x11 >= 2, do: tag_helper(:dice, @entry.x2) %>
+    <%= if @entry.x11 >= 3, do: tag_helper(:dice, @entry.x3) %>
+    <%= if @entry.x11 >= 4, do: tag_helper(:dice, @entry.x4) %>
+    <%= if @entry.x11 >= 5, do: tag_helper(:dice, @entry.x5) %>
+    <%= if @entry.x11 >= 6, do: tag_helper(:dice, @entry.x6) %>
+    <%= if @entry.x11 >= 7, do: tag_helper(:dice, @entry.x7) %>
+    <%= if @entry.x11 == 8, do: tag_helper(:dice, @entry.x8) %>
+    <%= if input_value(@f, :result), do: separator() %>
+  <% end %>
+  <%= if input_value(@f, :result), do: tag_helper(:text, @entry.x12) %>
+</div>
