+++ lib/dsa_web/templates/log/trial_result.html.leex
@@ -1,0 +1,9 @@
+<div class='flex'>
+  <%= if input_value(@f, :dice) do %>
+    <%= tag_helper :dice, @entry.x7 %>
+    <%= tag_helper :dice, @entry.x8 %>
+    <%= tag_helper :dice, @entry.x9 %>
+    <%= if input_value(@f, :result), do: separator() %>
+  <% end %>
+  <%= if input_value(@f, :result), do: result_tag(:talent, @entry.x12) %>
+</div>
