+++ lib/dsa_web/templates/log/unknown.html.leex
@@ -1,0 +1,1 @@
+Unbekannter Logeintrag
